架构优化
可同时在三种文本处理架构的视图间切换。

+++ CoreTextView_Lab0/Res/Utils/UIColor+Utils.h
@@ -12,6 +12,6 @@
 @interface UIColor (Utils)
 
 + (UIColor *)colorWithARGB:(NSUInteger)argb;
-+ (UIColor *)colorWithAlpha:(NSUInteger)alpha Red:(NSUInteger)red green:(NSUInteger)green blue:(NSUInteger)blue;
++ (UIColor *)colorWithAlpha:(CGFloat)alpha Red:(CGFloat)red green:(CGFloat)green blue:(CGFloat)blue;
 
 @end
