+++ CoreTextView_Lab0/Res/Utils/UIColor+Utils.m
@@ -22,7 +22,7 @@
 	return [UIColor colorWithRed:r / base green:g / base blue:b / base alpha:a / base];
 }
 
-+ (UIColor *)colorWithAlpha:(NSUInteger)alpha Red:(NSUInteger)red green:(NSUInteger)green blue:(NSUInteger)blue
++ (UIColor *)colorWithAlpha:(CGFloat)alpha Red:(CGFloat)red green:(CGFloat)green blue:(CGFloat)blue
 {
 	CGFloat base = 255.0;
 	return [UIColor colorWithRed:red / base green:green / base blue:blue / base alpha:alpha / base];
